+++ {{cookiecutter.directory_name}}/apps/api/decorators.py
@@ -1,13 +1,20 @@
-from functools import wraps
+from http import HTTPStatus
+
+from django.utils.translation import gettext as _
+
+from apps.api.errors import ProblemDetailException
 
 
-def require_apikey(view_func):
+def permission_required(perm):
     """
-    Require valid signature for specified endpoint
-    :param view_func:
-    :return:
+    Mark a view function to check specific user permission.
     """
-    def wrapped_view(*args, **kwargs):
-        return view_func(*args, **kwargs)
-    wrapped_view.require_apikey = True
-    return wraps(view_func)(wrapped_view)
+    def decorator(func):
+        def wrapper(request, *args, **kwargs):
+            if request.user.has_perm(perm):
+                return func(request, *args, **kwargs)
+            else:
+                raise ProblemDetailException(request, title=_('Permission denied.'), status_code=HTTPStatus.FORBIDDEN)
+
+        return wrapper
+    return decorator