+++ {{cookiecutter.directory_name}}/apps/core/models/user.py
@@ -1,5 +1,6 @@
 from django.contrib.auth.base_user import AbstractBaseUser
 from django.contrib.auth.models import PermissionsMixin
+from django.utils.translation import gettext as _
 from django.db import models
 
 from apps.core.managers.user import UserManager
@@ -14,10 +15,10 @@ class User(BaseModel, AbstractBaseUser, PermissionsMixin):
         default_permissions = ('add', 'delete')
 
     # Basic info
-    email = models.EmailField(unique=True)
-    name = models.CharField(max_length=30)
-    surname = models.CharField(max_length=150)
-    is_active = models.BooleanField(default=True)
+    email = models.EmailField(null=False, unique=True, verbose_name=_('user_email'))
+    name = models.CharField(null=False, max_length=30, verbose_name=_('user_name'))
+    surname = models.CharField(null=False, max_length=150, verbose_name=_('user_surname'))
+    is_active = models.BooleanField(null=False, default=True, verbose_name=_('user_is_active'))
 
     objects = UserManager()
     all_objects = UserManager(alive_only=False)